requirements.txt
loguru~=0.7.3
python-dotenv~=1.0.1
requests~=2.32.3
drissionpage~=4.1.0.17
watchdog==3.0.0
psycopg2-binary==2.9.9